+++ src/Checkout/ErrorController.php
@@ -7,8 +7,10 @@
 
 namespace Swag\PayPal\Checkout;
 
+use Psr\Log\LoggerInterface;
 use Shopware\Core\Framework\Routing\Annotation\RouteScope;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\Session\Session;
 use Symfony\Component\Routing\Annotation\Route;
@@ -19,20 +21,17 @@ use Symfony\Contracts\Translation\TranslatorInterface;
  */
 class ErrorController extends AbstractController
 {
-    /**
-     * @var Session
-     */
-    private $session;
+    private Session $session;
 
-    /**
-     * @var TranslatorInterface
-     */
-    private $translator;
+    private TranslatorInterface $translator;
 
-    public function __construct(Session $session, TranslatorInterface $translator)
+    private LoggerInterface $logger;
+
+    public function __construct(Session $session, TranslatorInterface $translator, LoggerInterface $logger)
     {
         $this->session = $session;
         $this->translator = $translator;
+        $this->logger = $logger;
     }
 
     /**
@@ -43,9 +42,10 @@ class ErrorController extends AbstractController
      *     defaults={"XmlHttpRequest"=true}
      * )
      */
-    public function addErrorMessage(): Response
+    public function addErrorMessage(Request $request): Response
     {
         $this->session->getFlashBag()->add('danger', $this->translator->trans('paypal.general.paymentError'));
+        $this->logger->notice('Storefront checkout error', ['error' => $request->get('error')]);
 
         return new Response('', Response::HTTP_NO_CONTENT);
     }