+++ src/Resources/app/storefront/src/smart-payment-buttons/swag-paypal.marks.js
@@ -10,6 +10,34 @@ export default class SwagPayPalMarks extends SwagPaypalAbstractButtons {
         clientId: '',
 
         /**
+         * This option specifies the language of the PayPal button
+         *
+         * @type string
+         */
+        languageIso: 'en_GB',
+
+        /**
+         * This options specifies the currency of the PayPal button
+         *
+         * @type string
+         */
+        currency: 'EUR',
+
+        /**
+         * This options defines the payment intent
+         *
+         * @type string
+         */
+        intent: 'capture',
+
+        /**
+         * This option toggles the PayNow/Login text at PayPal
+         *
+         * @type boolean
+         */
+        commit: true,
+
+        /**
          * This option toggles if credit card and ELV should be shown
          *
          * @type boolean