+++ src/Resources/app/storefront/src/swag-paypal.abstract-buttons.js
@@ -85,7 +85,7 @@ export default class SwagPaypalAbstractButtons extends Plugin {
         return config;
     }
 
-    createError() {
+    createError(error) {
         if (process.env.NODE_ENV !== 'production' && typeof console !== 'undefined' && typeof this._client === 'undefined') {
             console.error('No StoreApiClient defined in child plugin class');
             return;
@@ -101,7 +101,8 @@ export default class SwagPaypalAbstractButtons extends Plugin {
         }
 
         const requestPayload = {
-            _csrf_token: DomAccess.getDataAttribute(this.el, 'data-swag-pay-pal-add-error-token')
+            _csrf_token: DomAccess.getDataAttribute(this.el, 'data-swag-pay-pal-add-error-token'),
+            error: error
         };
 
         this._client.post(addErrorUrl, JSON.stringify(requestPayload), () => {